Added functionality
Ui behaviour is now appropriate, eg.

If a SID certificate combination does not exist, then next is not
available as there is no subscription to move forward with, but create
is available. Create provides the means to create their new selected
subscription. Once create has been clicked, next become available and
create unavailable. Next then loads the certificate saved and proceeds
with the wizard. Create should also save a copy of the the certificate
to the desktop for a user to upload to azure but does not yet.

This functionality is not adequately tested yet.